BackEnd/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.6
websocket-client==0.48.0
Werkzeug==0.16.0
